+++ src/charts/BarChart/BarChart_snapshot.spec.js
@@ -1,19 +1,22 @@
 /* eslint-disable */
 // This file was automatically generated
 import React from 'react';
-import { shallow } from 'enzyme';
+import 'jest-styled-components';
 import BarChart from './BarChart';
+
+
+jest.mock('../../charts/HighChart', () => () => <high-charts />);
 
 describe('BarChart Component', () => {
   describe('snapshots', () => {
     it('should match snapshot', () => {
     
-      const wrapper1 = shallow(
+      const wrapper1 = mount(
         <BarChart title="Category mix" data={require('../fixtures/bar-chart-category-mix.js').default} />
       );
       expect(wrapper1).toMatchSnapshot();
 
-      const wrapper2 = shallow(
+      const wrapper2 = mount(
         <BarChart
           title="Category mix"
           data={require('../fixtures/bar-chart-category-mix-many.js').default}
@@ -22,7 +25,7 @@ describe('BarChart Component', () => {
       );
       expect(wrapper2).toMatchSnapshot();
 
-      const wrapper3 = shallow(
+      const wrapper3 = mount(
         <BarChart
           orientation="vertical"
           title="Price point"