+++ src/components/TabBar/TabBar_snapshot.spec.js
@@ -1,14 +1,17 @@
 /* eslint-disable */
 // This file was automatically generated
 import React from 'react';
-import { shallow } from 'enzyme';
+import 'jest-styled-components';
 import TabBar from './TabBar';
+
+
+jest.mock('../../charts/HighChart', () => () => <high-charts />);
 
 describe('TabBar Component', () => {
   describe('snapshots', () => {
     it('should match snapshot', () => {
     
-      const wrapper1 = shallow(
+      const wrapper1 = mount(
         <TabBar>
           <TabBar.Item value="overview">Proposal overview</TabBar.Item>
           <TabBar.Item value="analytics">Analytics</TabBar.Item>
@@ -16,7 +19,7 @@ describe('TabBar Component', () => {
       );
       expect(wrapper1).toMatchSnapshot();
 
-      const wrapper2 = shallow(
+      const wrapper2 = mount(
         <TabBar size="large" onSelect={console.log} defaultValue="analytics">
           <TabBar.Item value="overview">Proposal overview</TabBar.Item>
           <TabBar.Item value="analytics">Analytics</TabBar.Item>
@@ -24,7 +27,7 @@ describe('TabBar Component', () => {
       );
       expect(wrapper2).toMatchSnapshot();
 
-      const wrapper3 = shallow(
+      const wrapper3 = mount(
         <TabBar align="left">
           <TabBar.Item value="overview">Proposal overview</TabBar.Item>
           <TabBar.Item value="analytics">Analytics</TabBar.Item>